+++ src/api/api.ts
@@ -7,13 +7,33 @@ import type {
   CreateCorpusData,
   ResourceInfoAllData,
   ResourceInfoOneData,
-  JobState,
-  JobType,
   ListExportsData,
   AdminModeStatusData,
   CreateMetadataData,
   ProgressHandler,
+  JobStateMap,
 } from "@/api/api.types";
+
+/** Create a `text/yaml` file object with content */
+const yamlAsFile = (filename: string, yaml: string): File =>
+  new File([yaml], filename, { type: "text/yaml" });
+
+/** Create a form data object with one or more files under `"files[]"` */
+function filesFormData(...files: File[]): FormData {
+  const formData = new FormData();
+  files.forEach((file) => formData.append("files[]", file));
+  return formData;
+}
+
+/** Handle an exception from an API call that may be encoded as Blob */
+async function rethrowBlobError(error: any): Promise<never> {
+  if (error.response?.data instanceof Blob) {
+    // Parse JSON and replace the blob
+    const text = await error.response.data.text();
+    error.response.data = JSON.parse(text) as MinkResponse;
+  }
+  throw error;
+}
 
 /** Mink backend API client */
 class MinkApi {
@@ -37,20 +57,20 @@ class MinkApi {
     this.axios.defaults.headers["Authorization"] = jwt ? `Bearer ${jwt}` : null;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Documentation/operation/APIinfo */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Documentation/operation/APIinfo */
   async getInfo() {
     const response = await this.axios.get<MinkResponse<InfoData>>("info");
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Corpora/operation/listcorpora */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Corpora/operation/listcorpora */
   async listCorpora() {
     const response =
       await this.axios.get<MinkResponse<ListCorporaData>>("list-corpora");
     return response.data.corpora;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Corpora/operation/createcorpus */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Corpora/operation/createcorpus */
   async createCorpus() {
     const response =
       await this.axios.post<MinkResponse<CreateCorpusData>>("create-corpus");
@@ -66,7 +86,7 @@ class MinkApi {
     return response.data.resource_id;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Corpora/operation/removecorpus */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Corpora/operation/removecorpus */
   async removeCorpus(corpusId: string) {
     const response = await this.axios.delete<MinkResponse>("remove-corpus", {
       params: { corpus_id: corpusId },
@@ -74,7 +94,7 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Metadata/operation/removemetadata */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Metadata/operation/removemetadata */
   async removeMetadata(resourceId: string) {
     const response = await this.axios.delete<MinkResponse>("remove-metadata", {
       params: { corpus_id: resourceId },
@@ -82,11 +102,9 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Config/operation/uploadconfig */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Config/operation/uploadconfig */
   async uploadConfig(corpusId: string, config: string) {
-    const configFile = new File([config], "config.yaml", { type: "text/yaml" });
-    const formData = new FormData();
-    formData.append("files[]", configFile);
+    const formData = filesFormData(yamlAsFile("config.yaml", config));
     const response = await this.axios.put<MinkResponse>(
       "upload-config",
       formData,
@@ -95,16 +113,22 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Sources/operation/downloadsources */
-  async downloadSources(corpusId: string, filename: string, binary = false) {
-    const response = await this.axios.get<string>("download-sources", {
-      params: { corpus_id: corpusId, file: filename, zip: false },
-      responseType: binary ? "arraybuffer" : "text",
-    });
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Sources/operation/downloadsources */
+  async downloadSources<B extends boolean>(
+    corpusId: string,
+    filename: string,
+    binary: B,
+  ) {
+    const response = await this.axios
+      .get<B extends true ? Blob : string>("download-sources", {
+        params: { corpus_id: corpusId, file: filename, zip: false },
+        responseType: binary ? "blob" : "text",
+      })
+      .catch(rethrowBlobError);
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Exports/operation/downloadsourcetext */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Exports/operation/downloadsourcetext */
   async downloadSourceText(corpusId: string, filename: string) {
     const response = await this.axios.get<string>("download-source-text", {
       params: { corpus_id: corpusId, file: filename },
@@ -112,14 +136,13 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Sources/operation/uploadsources */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Sources/operation/uploadsources */
   async uploadSources(
     corpusId: string,
     files: File[],
     onProgress?: ProgressHandler,
   ) {
-    const formData = new FormData();
-    files.forEach((file) => formData.append("files[]", file));
+    const formData = filesFormData(...files);
     const response = await this.axios.put<MinkResponse>(
       "upload-sources",
       formData,
@@ -131,7 +154,7 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Sources/operation/removesources */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Sources/operation/removesources */
   async removeSource(corpusId: string, name: string) {
     const response = await this.axios.delete<MinkResponse>("remove-sources", {
       params: { corpus_id: corpusId, remove: name },
@@ -139,7 +162,7 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Config/operation/downloadconfig */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Config/operation/downloadconfig */
   async downloadConfig(corpusId: string) {
     const response = await this.axios.get<string>("download-config", {
       params: { corpus_id: corpusId },
@@ -147,11 +170,9 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Metadata/operation/uploadmetadatayaml */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Metadata/operation/uploadmetadatayaml */
   async uploadMetadataYaml(resourceId: string, yaml: string) {
-    const file = new File([yaml], "metadata.yaml", { type: "text/yaml" });
-    const formData = new FormData();
-    formData.append("files[]", file);
+    const formData = filesFormData(yamlAsFile("metadata.yaml", yaml));
     const response = await this.axios.put<MinkResponse>(
       "upload-metadata-yaml",
       formData,
@@ -160,7 +181,7 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Metadata/operation/downloadmetadatayaml */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Metadata/operation/downloadmetadatayaml */
   async downloadMetaataYaml(resourceId: string) {
     const response = await this.axios.get<string>("download-metadata-yaml", {
       params: { corpus_id: resourceId },
@@ -168,14 +189,14 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Process-Corpus/operation/resourceinfo */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Process-Corpus/operation/resourceinfo */
   async resourceInfoAll() {
     const response =
       await this.axios.get<MinkResponse<ResourceInfoAllData>>("resource-info");
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Process-Corpus/operation/resourceinfo */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Process-Corpus/operation/resourceinfo */
   async resourceInfoOne(corpusId: string) {
     const response = await this.axios.get<MinkResponse<ResourceInfoOneData>>(
       "resource-info",
@@ -184,7 +205,7 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Process-Corpus/operation/runSparv */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Process-Corpus/operation/runSparv */
   async runSparv(corpusId: string) {
     const response = await this.axios
       .put<MinkResponse<ResourceInfoOneData>>("run-sparv", null, {
@@ -196,17 +217,17 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Process-Corpus/operation/abortjob */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Process-Corpus/operation/abortjob */
   async abortJob(corpusId: string) {
-    const response = await this.axios.post<
-      MinkResponse<Record<JobType, JobState>>
-    >("abort-job", null, {
-      params: { corpus_id: corpusId },
-    });
+    const response = await this.axios.post<MinkResponse<JobStateMap>>(
+      "abort-job",
+      null,
+      { params: { corpus_id: corpusId } },
+    );
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Exports/operation/listexports */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Exports/operation/listexports */
   async listExports(corpusId: string) {
     const response = await this.axios.get<MinkResponse<ListExportsData>>(
       "list-exports",
@@ -215,25 +236,29 @@ class MinkApi {
     return response.data.contents;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Exports/operation/downloadexports */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Exports/operation/downloadexports */
   async downloadExports(corpusId: string) {
-    const response = await this.axios.get<Blob>("download-exports", {
-      params: { corpus_id: corpusId },
-      responseType: "blob",
-    });
+    const response = await this.axios
+      .get<Blob>("download-exports", {
+        params: { corpus_id: corpusId },
+        responseType: "blob",
+      })
+      .catch(rethrowBlobError);
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Manage-Exports/operation/downloadexports */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Manage-Exports/operation/downloadexports */
   async downloadExportFile(corpusId: string, path: string) {
-    const response = await this.axios.get<Blob>("download-exports", {
-      params: { corpus_id: corpusId, file: path, zip: false },
-      responseType: "blob",
-    });
+    const response = await this.axios
+      .get<Blob>("download-exports", {
+        params: { corpus_id: corpusId, file: path, zip: false },
+        responseType: "blob",
+      })
+      .catch(rethrowBlobError);
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Process-Corpus/operation/installinKorp */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Process-Corpus/operation/installinKorp */
   async installKorp(corpusId: string) {
     const response = await this.axios.put<MinkResponse<ResourceInfoOneData>>(
       "install-korp",
@@ -243,7 +268,7 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Process-Corpus/operation/installinStrix */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Process-Corpus/operation/installinStrix */
   async installStrix(corpusId: string) {
     const response = await this.axios.put<MinkResponse<ResourceInfoOneData>>(
       "install-strix",
@@ -253,7 +278,7 @@ class MinkApi {
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Admin-Mode/operation/adminmodestatus */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Admin-Mode/operation/adminmodestatus */
   async adminModeStatus() {
     const response =
       await this.axios.get<MinkResponse<AdminModeStatusData>>(
@@ -262,13 +287,13 @@ class MinkApi {
     return response.data.admin_mode_status;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Admin-Mode/operation/adminmodeon */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Admin-Mode/operation/adminmodeon */
   async adminModeOn() {
     const response = await this.axios.post<MinkResponse>("admin-mode-on");
     return response.data;
   }
 
-  /** @see https://ws.spraakbanken.gu.se/ws/mink/api-doc#tag/Admin-Mode/operation/adminmodeoff */
+  /** @see https://ws.spraakbanken.gu.se/docs/mink#tag/Admin-Mode/operation/adminmodeoff */
   async adminModeOff() {
     const response = await this.axios.post<MinkResponse>("admin-mode-off");
     return response.data;