+++ src/components/MaxHeight.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { ref } from "vue";
 import { useElementSize, useToggle } from "@vueuse/core";
+import { PhCaretDoubleDown, PhCaretDoubleUp } from "@phosphor-icons/vue";
 import ActionButton from "./ActionButton.vue";
 
 defineProps<{
@@ -31,7 +32,8 @@ const [expanded, toggleExpanded] = useToggle();
 
     <div v-if="height > maxHeight" class="text-center p-2 text-sm">
       <ActionButton class="button-slim" @click="toggleExpanded()">
-        <icon :icon="expanded ? 'angles-up' : 'angles-down'" />
+        <PhCaretDoubleUp v-if="expanded" class="inline mb-1 mr-0.5" />
+        <PhCaretDoubleDown v-else class="inline mb-1 mr-0.5" />
         {{ expanded ? $t("expand.close") : $t("expand.open") }}
       </ActionButton>
     </div>