+++ src/corpus/config/CorpusConfigCustom.vue
@@ -1,5 +1,6 @@
 <script setup lang="ts">
 import type { AxiosError } from "axios";
+import { PhWarning } from "@phosphor-icons/vue";
 import useCorpusIdParam from "@/corpus/corpusIdParam.composable";
 import useConfig from "@/corpus/config/config.composable";
 import { useAuth } from "@/auth/auth.composable";
@@ -9,6 +10,7 @@ import LayoutBox from "@/components/LayoutBox.vue";
 import type { MinkResponse } from "@/api/api.types";
 import useMessenger from "@/message/messenger.composable";
 import SyntaxHighlight from "@/components/SyntaxHighlight.vue";
+import PendingContent from "@/spin/PendingContent.vue";
 
 const corpusId = useCorpusIdParam();
 const { config, uploadConfigRaw } = useConfig(corpusId);
@@ -46,12 +48,23 @@ async function upload(files: File[]) {
 
   <div class="flex flex-wrap gap-4">
     <LayoutBox class="w-96 grow" :title="$t('show')">
-      <SyntaxHighlight v-if="config" language="yaml" :code="config" />
+      <PendingContent :on="`corpus/${corpusId}/config`">
+        <SyntaxHighlight v-if="config" language="yaml" :code="config" />
+      </PendingContent>
     </LayoutBox>
 
     <LayoutBox class="w-96 grow" :title="$t('upload')">
-      <HelpBox>{{ $t("config.custom.upload.help") }}</HelpBox>
-      <FileUpload :file-handler="upload" accept=".yaml,.yml" />
+      <HelpBox important>
+        <PhWarning class="inline mb-1 mr-1" />
+        {{ $t("config.custom.upload.caution") }}
+      </HelpBox>
+      <HelpBox important>
+        <PhWarning class="inline mb-1 mr-1" />
+        {{ $t("config.custom.upload.overwrite") }}
+      </HelpBox>
+      <PendingContent :on="`corpus/${corpusId}/config`" blocking>
+        <FileUpload :file-handler="upload" accept=".yaml,.yml" />
+      </PendingContent>
     </LayoutBox>
   </div>
 </template>