+++ src/corpus/explore/ToolPanel.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { PhArrowSquareOut, PhGearFine } from "@phosphor-icons/vue";
 import ActionButton from "@/components/ActionButton.vue";
 import UrlButton from "@/components/UrlButton.vue";
 
@@ -28,7 +29,7 @@ defineEmits<{
           :class="{ 'button-primary': canInstall && !isInstalled }"
           @click="$emit('install')"
         >
-          <icon :icon="['fas', 'gears']" class="mr-1" />
+          <PhGearFine weight="bold" class="inline mb-1 mr-1" />
           {{
             $t(
               !isInstalled
@@ -44,7 +45,7 @@ defineEmits<{
           target="_blank"
           class="button-primary"
         >
-          <icon icon="up-right-from-square" size="sm" class="mr-1" />
+          <PhArrowSquareOut weight="bold" class="inline mb-1 mr-1" />
           {{ $t("exports.tools.view") }}
         </UrlButton>
       </div>