+++ src/corpus/job/JobStatus.vue
@@ -7,7 +7,6 @@ import PendingContent from "@/spin/PendingContent.vue";
 import useCorpusIdParam from "@/corpus/corpusIdParam.composable";
 import { useCorpusState } from "@/corpus/corpusState.composable";
 import ActionButton from "@/components/ActionButton.vue";
-import TerminalOutput from "@/components/TerminalOutput.vue";
 import ProgressBar from "@/components/ProgressBar.vue";
 import TextData from "@/components/TextData.vue";
 
@@ -99,10 +98,3 @@ const hasStarted = computed(
     </table>
   </PendingContent>
 </template>
-
-<style scoped>
-/* Override max-height when user is resizing. */
-.resize-y[style*="height"] {
-  max-height: unset;
-}
-</style>