+++ src/i18n/LocaleSwitcher.vue
@@ -1,10 +1,11 @@
 <script setup lang="ts">
+import { PhGlobe } from "@phosphor-icons/vue";
 import { languageNames, type LocaleId } from "./i18n";
 </script>
 
 <template>
   <div>
-    <icon icon="language" />
+    <PhGlobe class="inline mb-0.5" />
     <select
       v-model="$i18n.locale"
       class="bg-transparent border-0 appearance-none"